Add folder structure and ts page demo
The demo is mainly for me, but others are welcome to use it for reference as well. It shows how to use component-level styling instead of global styling.

Feel free to edit/remove what is currently in the characters page, again mainly keeping it there for my own reference since it's my first time using a lot of these tools and frameworks.